WIP: modularize the metadata checker class
