+++ pydatacuration/checker/metadata_checker.py
@@ -1,22 +1,36 @@
 """MetadataChecker class for checking dataset metadata."""
 
 import jmespath
+from loguru import logger
+
+from pydatacuration.checker.check_result_writer import CheckResultWriter
 
 
 class MetadataChecker:
     """Class for checking dataset metadata."""
 
-    def __init__(self, metadata: dict) -> None:
+    def __init__(self, metadata: dict, checklist_result_writer: CheckResultWriter) -> None:
         """Initialize the MetadataChecker class.
 
         Args:
             metadata (dict): Metadata from the JSON file.
+            checklist_result_writer (CheckResultWriter): Writer for storing check results.
         """
         self.metadata = metadata
+        self.checklist_result_writer = checklist_result_writer
 
-    def _read_metadata_cm_field(self, field: str, subfield=None):
-        # TODO: fix the logic of subfield
+        self.author_info_dict = self.get_author_cm_field()
 
+    def _get_metadata_cm_field(self, field: str, subfield: str | None = None) -> str | None:
+        """Get the value of a metadata field from the metadata JSON file.
+
+        Args:
+            field (str): Metadata field to check.
+            subfield (str, optional): Subfield to check. Defaults to None.
+
+        Returns:
+            result (str | None): Value of the metadata field or None if it doesn't exist
+        """
         if subfield:
             query_string = (
                 f'data.latestVersion.metadataBlocks.citation.fields[?typeName==`{field}`].value[].[{subfield}][].value'  # noqa: E501
@@ -27,8 +41,8 @@ class MetadataChecker:
             result = jmespath.search(query_string, self.metadata)
         return result
 
-    def check_metadata_cm_field(self, field: str) -> tuple[str | None, bool]:
-        r"""Check if a metadata field exists in the metadata JSON file.
+    def get_metadata_cm_field(self, field: str) -> tuple[str | None, bool]:
+        r"""Get the value of a metadata field from the metadata JSON file.
 
         Args:
             field (str): Metadata field to check.\n
@@ -44,14 +58,14 @@ class MetadataChecker:
         else:
             subfield = None
 
-        result = self._read_metadata_cm_field(field, subfield)
+        result = self._get_metadata_cm_field(field, subfield)
         if result:
             return result, True
 
         return None, False
 
-    def check_author_cm_field(self) -> list[dict]:
-        r"""Check if the author metadata fields exist in the metadata JSON file.
+    def get_author_cm_field(self) -> list[dict]:
+        r"""Get the author metadata fields from the metadata JSON file.
 
         Returns:
             result (list[dict]): List of dictionaries containing author metadata fields
@@ -60,3 +74,94 @@ class MetadataChecker:
         result = jmespath.search(query_string, self.metadata)
 
         return result or []
+
+    def check_missing_required_fields(self) -> None:
+        """Check for missing required metadata fields."""
+        missing_required_fields = []
+
+        required_fields = ['title', 'dsDescription', 'subject']
+
+        for field in required_fields:
+            return_value = self.get_metadata_cm_field(field)
+            if return_value[1] is False:
+                logger.info(f'Missing metadata found in the {field}')
+                missing_required_fields.append(field)
+
+        self.checklist_result_writer.write(
+            check_id='missing_required_fields',
+            check_name='Missing Required Metadata Fields',
+            description='Required metadata fields that are empty or missing',
+            unit='field',
+            results=missing_required_fields,
+        )
+
+    def check_missing_author_affiliation(self) -> None:
+        """Check for missing author affiliation."""
+        authors_missing_affiliation = []
+
+        for item in self.author_info_dict:
+            if item.get('authorAffiliation') is None:
+                author_name = item.get('authorName')
+                logger.info(f'Missing metadata found in authorAffiliation field for author: {author_name}')
+                authors_missing_affiliation.append(author_name)
+
+        self.checklist_result_writer.write(
+            check_id='authors_missing_affiliation',
+            check_name='Author affiliation field',
+            description='Authors missing institutional affiliation information',
+            unit='author',
+            results=authors_missing_affiliation,
+        )
+
+    def check_missing_author_identifier(self) -> None:
+        """Check for missing author identifier."""
+        authors_missing_identifier = []
+
+        for item in self.author_info_dict:
+            if item.get('authorIdentifier') is None:
+                author_name = item.get('authorName')
+                logger.info(f'Missing metadata found in authorIdentifier field for author: {author_name}')
+                authors_missing_identifier.append(author_name)
+
+        self.checklist_result_writer.write(
+            check_id='authors_missing_identifier',
+            check_name='Author Research ID field',
+            description='Authors missing personal identifier (ORCID, etc.)',
+            unit='author',
+            results=authors_missing_identifier,
+        )
+
+    def check_missing_author_identifier_scheme(self) -> None:
+        """Check for missing author identifier scheme."""
+        authors_missing_scheme = []
+
+        for item in self.author_info_dict:
+            if item.get('authorIdentifierScheme') is None:
+                author_name = item.get('authorName')
+                logger.info(f'Missing metadata found in authorIdentifierScheme field for author: {author_name}')
+                authors_missing_scheme.append(author_name)
+
+        self.checklist_result_writer.write(
+            check_id='authors_missing_scheme',
+            check_name='Authors Research Identifier Scheme',
+            description='Authors missing identifier scheme information',
+            unit='author',
+            results=authors_missing_scheme,
+        )
+
+    def check_missing_institutional_affiliation(self) -> None:
+        """Check for missing institutional affiliation."""
+        # Check if at least one author has affiliation with 'University of Toronto' (Non-case sensitive)
+        ut_variants = ['university of toronto', 'uoft', 'u of t']
+        author_affiliation_ut_num = len(
+            [
+                item
+                for item in self.author_info_dict
+                if item.get('authorAffiliation') is not None
+                and any(variant in item.get('authorAffiliation', '').lower() for variant in ut_variants)
+            ]
+        )  # noqa: E501
+        if author_affiliation_ut_num == 0:
+            logger.info('None of the authors have listed affiliation with University of Toronto')
+
+        # TODO: Write this to db and make it a check in the checklist
